refactor/NFS move file_metadata
